+++ python/metatensor-learn/CHANGELOG.md
@@ -8,7 +8,7 @@ follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Added `metatensor.learn.nn.EquivariantTransform` to apply any 
+- Added `metatensor.learn.nn.EquivariantTransformation` to apply any 
   `torch.nn.Module` to invariants computed from the norm over components of covariant 
   blocks. The transformed invariants are then elementwise multiplied back to the 
   covariant blocks. For invariant blocks, the `torch.nn.Module` is applied as is (#744)